put queryParams into context instead
